checks if hard constraints are sat.added testing class

+++ flask-scheduling-server/Scheduler.py
@@ -164,6 +164,7 @@ def startAlgo():
     elasped = time.time() - start
     print("time taken = {}s".format(elasped))
     answer = sorted(lss,key = lambda coord : coord.fitnessValue)[0]
+    if answer.fitnessValue > 1: raise Exception("Hard Constraints not Satisfied")
     #key into database
     #deletes all previous data
     models.CourseClass.query.delete()
@@ -188,10 +189,13 @@ def startAlgo():
         db.session.add(courseClassDb)
 
 
-startAlgo()
+if __name__ == "__main__":
+    if not models.existDB():
+        models.createDB()
+    startAlgo()
 
-for cc in models.CourseClass.query.all():
-    print(cc)
+    for cc in models.CourseClass.query.all():
+        print(cc)
 
 #TODO 
 #1 DONE:implement lecture all student groups all profs
